pyproject.toml
[project]
name = "aiagent"
version = "0.1.0"
description = "Эталонный шаблон AI-агента с RAG, LangGraph и Google Sheets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- API ---
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sse-starlette>=2.1",         # SSE streaming для Vue 3 EventSource
    "python-multipart>=0.0.20",   # Загрузка файлов

    # --- LangChain / LangGraph ---
    "langchain>=0.3",
    "langchain-openai>=0.3",              # OpenRouter совместим с OpenAI API
    "langgraph>=0.3",
    "langgraph-checkpoint-sqlite>=2.0",   # Персистентная память через SQLite

    # --- RAG / Vector Store ---
    "langchain-chroma>=0.2",
    "chromadb>=0.6",

    # --- Загрузчики документов ---
    "langchain-community>=0.3",           # BaseLoader, DirectoryLoader и др.
    "pypdf>=5.0",                         # PDF
    "python-docx>=1.1",                   # DOCX (альтернативный загрузчик)
    "docx2txt>=0.8",                      # DOCX (Docx2txtLoader)
    "unstructured[html,csv]>=0.16",       # HTML, CSV

    # --- Google Sheets ---
    "gspread>=6.1",
    "google-auth>=2.35",

    # --- Конфигурация / Валидация ---
    "pydantic>=2.10",
    "pydantic-settings>=2.6",

    # --- Логирование ---
    "loguru>=0.7",

    # --- HTTP клиент ---
    "httpx>=0.28",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "httpx",                # AsyncClient для тестов
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# --- Ruff ---
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = ["B008"]  # Вызовы функций в аргументах FastAPI Depends — это норма

[tool.ruff.lint.isort]
known-first-party = ["app"]

# --- MyPy ---
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# --- Pytest ---
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
